+++ spec/support/shared_contexts/this_api.rb
@@ -18,13 +18,14 @@ shared_context 'this api' do
         end
 
         class Nested < Grape::Entity
-          expose :attr, documentation: { type: 'string', desc: 'Attribute' }
+          expose :attr, documentation: { required: true, type: 'string', desc: 'Attribute' }
           expose :nested_attrs, merge: true, using: ThisApi::Entities::Error
         end
 
         class Something < Grape::Entity
           expose :text, documentation: { type: 'string', desc: 'Content of something.' }
           expose :colors, documentation: { type: 'string', desc: 'Colors', is_array: true }
+          expose :hidden_attr, documentation: { type: 'string', desc: 'Hidden', hidden: true }
           expose :kind, using: Kind, documentation: { type: 'ThisApi::Kind', desc: 'The kind of this something.' }
           expose :kind2, using: Kind, documentation: { desc: 'Secondary kind.' }
           expose :kind3, using: ThisApi::Entities::Kind, documentation: { desc: 'Tertiary kind.' }